MR - fixed mcphasit ion indexing bug.
The bug in module mcphas is only important for the case when all the
following conditions are fulfilled, this is why it did not turn up
earlier:
- more than 1 atom in crystal basis, which are not ferromagnetically
  aligned
- magnetic structure supercell is larger than primitive crystal lattice
  unit cell

Output of free energies, magnetisation and magnetic energy is still
correct. However, magnetic structure and meanfields are wrong and if
used in mcdisp or mcdiff will lead to wrong results.