Added routines to calculate energy levels up to a limit in ic1ion
Added new spectre_eig() function in spectre.cpp - using a method similar
to that used in the XTAL84 programs. The Coulomb+SpinOrbit Hamiltonian
is first diagonalised, and its eigenvector matrix used to transform
the full IC Hamiltonian. Then states with energy above a limit is
truncated, and the truncated Hamiltonian diagonalised to get the energies.
The eigenvectors of the transformed Hic is converted back into the
original basis using the eigenvectors of Hc+Hso.